Move unit tests to live in respective modules
