+++ crates/module-system/module-implementations/sov-evm/src/evm/executor.rs
@@ -68,3 +68,42 @@ pub(crate) fn inspect<DB: Database<Error = Infallible> + DatabaseCommit>(
 
     evm.inspect_tx(tx)
 }
+
+#[cfg(test)]
+mod tests {
+    use alloy_primitives::U256;
+    use revm::primitives::hardfork::SpecId;
+    use sov_modules_api::macros::config_value;
+
+    use super::*;
+
+    #[test]
+    fn cfg_test() {
+        let block_env = BlockEnv {
+            number: U256::from(10),
+            ..Default::default()
+        };
+
+        let cfg = EvmRuntimeConfig {
+            chain_spec: crate::EvmChainSpec {
+                limit_contract_code_size: Some(100),
+                ..Default::default()
+            },
+            hardforks: vec![(0, SpecId::SHANGHAI)],
+        };
+
+        let mut template_cfg_env = CfgEnv::default();
+        template_cfg_env.chain_id = 2;
+        template_cfg_env.disable_base_fee = true;
+
+        let cfg_env = get_cfg_env(&block_env, cfg, Some(template_cfg_env));
+
+        let mut expected_cfg_env = CfgEnv::default();
+        expected_cfg_env.chain_id = config_value!("CHAIN_ID");
+        expected_cfg_env.disable_base_fee = true;
+        expected_cfg_env.limit_contract_code_size = Some(100);
+        expected_cfg_env.spec = SpecId::SHANGHAI;
+
+        assert_eq!(expected_cfg_env, cfg_env);
+    }
+}
