+++ crates/module-system/module-implementations/sov-evm/src/tests/mod.rs
@@ -1,2 +1,1 @@
-mod cfg_tests;
 mod tx_tests;
